Moving the self doc config to a new directory
