+++ .gitignore
@@ -1,7 +1,6 @@
 composer.phar
 /vendor/
 .idea
-/SelfDoc/__cache/
 /demo/demo1/docs/
 /demo/demo1/__cache/
 /demo/demo2-templates-generation/__cache/
