--- SelfDoc/Configuration/config.yaml
@@ -1,25 +1,0 @@
-project_root: '%DOC_GEN_LIB_PATH%'
-templates_dir: '%DOC_GEN_LIB_PATH%/SelfDoc/Configuration/templates'
-cache_dir: '%project_root%/SelfDoc/__cache'
-language_handlers:
-  php:
-    class: \BumbleDocGen\LanguageHandler\Php\PhpHandler
-    settings:
-      file_source_base_url: 'https://github.com/bumble-tech/bumble-doc-gen/blob/master'
-      async_source_loading_enabled: true
-source_locators:
-  - class: \BumbleDocGen\Core\Parser\SourceLocator\RecursiveDirectoriesSourceLocator
-    arguments:
-      directories:
-        - "%project_root%/src"
-        - "%project_root%/SelfDoc"
-twig_filters:
-  - class: \SelfDoc\Configuration\Twig\CustomFilter\EvalString
-twig_functions:
-  - class: \SelfDoc\Configuration\Twig\CustomFunction\FindEntitiesClassesByCollectionClassName
-  - class: \SelfDoc\Configuration\Twig\CustomFunction\PrintClassCollectionAsGroupedTable
-  - class: \SelfDoc\Configuration\Twig\CustomFunction\GetConfigParametersDescription
-plugins:
-  - class: \SelfDoc\Configuration\Plugin\RoaveStubber\BetterReflectionStubberPlugin
-  - class: \SelfDoc\Configuration\Plugin\TwigFilterClassParser\TwigFilterClassParserPlugin
-  - class: \SelfDoc\Configuration\Plugin\TwigFunctionClassParser\TwigFunctionClassParserPlugin
