--- SelfDoc/docGen.php
@@ -1,8 +1,0 @@
-#!/usr/bin/env php
-<?php
-
-ini_set('memory_limit', '-1');
-
-require_once __DIR__ . '/../vendor/autoload.php';
-
-(new \SelfDoc\Console\App())->run();
