+++ selfdoc/Plugin/TwigFilterClassParser/FilterClassPluginTwigEnvironment.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SelfDoc\Configuration\Plugin\TwigFunctionClassParser;
+namespace SelfDocConfig\Plugin\TwigFilterClassParser;
 
 use Twig\Environment;
 use Twig\Error\LoaderError;
@@ -10,7 +10,7 @@ use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
 use Twig\Loader\FilesystemLoader;
 
-final class FunctionClassPluginTwigEnvironment
+final class FilterClassPluginTwigEnvironment
 {
     private Environment $twig;
