+++ selfdoc/Plugin/TwigFilterClassParser/TwigFilterClassParserPlugin.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace SelfDoc\Configuration\Plugin\TwigFunctionClassParser;
+namespace SelfDocConfig\Plugin\TwigFilterClassParser;
 
 use BumbleDocGen\Core\Configuration\Configuration;
 use BumbleDocGen\Core\Configuration\Exception\InvalidConfigurationParameterException;
 use BumbleDocGen\Core\Plugin\Event\Renderer\OnLoadEntityDocPluginContent;
 use BumbleDocGen\Core\Plugin\PluginInterface;
 use BumbleDocGen\Core\Renderer\Context\RendererContext;
-use BumbleDocGen\Core\Renderer\Twig\Function\CustomFunctionInterface;
+use BumbleDocGen\Core\Renderer\Twig\Filter\CustomFilterInterface;
 use BumbleDocGen\LanguageHandler\Php\Parser\Entity\ClassEntity;
 use BumbleDocGen\LanguageHandler\Php\Parser\Entity\ClassEntityCollection;
 use BumbleDocGen\LanguageHandler\Php\Parser\Entity\Exception\ReflectionException;
@@ -18,16 +18,16 @@ use BumbleDocGen\LanguageHandler\Php\Renderer\EntityDocRenderer\PhpClassToMd\Php
 use DI\DependencyException;
 use DI\NotFoundException;
 
-final class TwigFunctionClassParserPlugin implements PluginInterface
+final class TwigFilterClassParserPlugin implements PluginInterface
 {
-    private const TWIG_FUNCTION_DIR_NAMES = [
-        '/src/Core/Renderer/Twig/Function',
-        '/src/LanguageHandler/Php/Renderer/Twig/Function'
+    private const TWIG_FILTER_DIR_NAMES = [
+        '/src/Core/Renderer/Twig/Filter',
+        '/src/LanguageHandler/Php/Renderer/Twig/Filter'
     ];
-    public const PLUGIN_KEY = 'twigFunctionClassParserPlugin';
+    public const PLUGIN_KEY = 'twigFilterClassParserPlugin';
 
     public function __construct(
-        private FunctionClassPluginTwigEnvironment $twigEnvironment,
+        private FilterClassPluginTwigEnvironment $twigEnvironment,
         private RendererContext $context,
         private Configuration $configuration,
     ) {
@@ -47,17 +47,19 @@ final class TwigFunctionClassParserPlugin implements PluginInterface
      */
     public function onLoadEntityDocPluginContentEvent(OnLoadEntityDocPluginContent $event): void
     {
-        if ($event->getBlockType() !== PhpClassToMdDocRenderer::BLOCK_AFTER_MAIN_INFO) {
+        if (
+            $event->getBlockType() !== PhpClassToMdDocRenderer::BLOCK_AFTER_MAIN_INFO
+        ) {
             return;
         }
 
         $entity = $event->getEntity();
-        if (!is_a($entity, ClassEntity::class) || !$this->isCustomTwigFunction($event->getEntity())) {
+        if (!is_a($entity, ClassEntity::class) || !$this->isCustomTwigFilter($event->getEntity())) {
             return;
         }
 
         try {
-            $pluginResult = $this->twigEnvironment->render('twigFunctionInfoBlock.twig', [
+            $pluginResult = $this->twigEnvironment->render('twigFilterInfoBlock.twig', [
                 'classEntity' => $entity,
             ]);
         } catch (\Exception) {
@@ -69,17 +71,17 @@ final class TwigFunctionClassParserPlugin implements PluginInterface
 
     /**
      * @throws NotFoundException
-     * @throws DependencyException
      * @throws ReflectionException
+     * @throws DependencyException
      * @throws InvalidConfigurationParameterException
      */
     public function afterLoadingClassEntityCollection(AfterLoadingClassEntityCollection $event): void
     {
         foreach ($event->getClassEntityCollection() as $classEntity) {
-            if ($this->isCustomTwigFunction($classEntity) && $classEntity->isInstantiable()) {
+            if ($this->isCustomTwigFilter($classEntity)) {
                 $classEntity->loadPluginData(
                     self::PLUGIN_KEY,
-                    $this->getFunctionData($event->getClassEntityCollection(), $classEntity->getName()) ?? []
+                    $this->getFilterData($event->getClassEntityCollection(), $classEntity->getName()) ?? []
                 );
             }
         }
@@ -89,10 +91,10 @@ final class TwigFunctionClassParserPlugin implements PluginInterface
      * @throws ReflectionException
      * @throws InvalidConfigurationParameterException
      */
-    private function isCustomTwigFunction(ClassEntity $classEntity): bool
+    private function isCustomTwigFilter(ClassEntity $classEntity): bool
     {
-        foreach (self::TWIG_FUNCTION_DIR_NAMES as $dirName) {
-            if ($classEntity->implementsInterface(CustomFunctionInterface::class) && str_starts_with($classEntity->getFileName(), $dirName)) {
+        foreach (self::TWIG_FILTER_DIR_NAMES as $dirName) {
+            if (str_starts_with($classEntity->getFileName(), $dirName) && $classEntity->implementsInterface(CustomFilterInterface::class)) {
                 return true;
             }
         }
@@ -105,46 +107,43 @@ final class TwigFunctionClassParserPlugin implements PluginInterface
      * @throws ReflectionException
      * @throws InvalidConfigurationParameterException
      */
-    private function getAllUsedFunctions(): array
+    private function getAllUsedFilters(): array
     {
-        static $functions = null;
-        if (is_null($functions)) {
-            $functions = [];
-            $twigFunctions = iterator_to_array($this->configuration->getTwigFunctions());
+        static $filters = null;
+        if (is_null($filters)) {
+            $filters = [];
+            $twigFilters = iterator_to_array($this->configuration->getTwigFilters());
             foreach ($this->configuration->getLanguageHandlersCollection() as $languageHandler) {
-                $twigFunctions = array_merge($twigFunctions, iterator_to_array($languageHandler->getCustomTwigFunctions($this->context)));
+                $twigFilters = array_merge($twigFilters, iterator_to_array($languageHandler->getCustomTwigFilters($this->context)));
             }
-            foreach ($twigFunctions as $function) {
-                $functions[$function::class] = $function::getName();
+            foreach ($twigFilters as $filter) {
+                $filters[$filter::class] = $filter::getName();
             }
         }
-        return $functions;
+        return $filters;
     }
 
     /**
-     * @throws ReflectionException
      * @throws DependencyException
+     * @throws ReflectionException
      * @throws NotFoundException
      * @throws InvalidConfigurationParameterException
      */
-    private function getFunctionData(ClassEntityCollection $classEntityCollection, string $className): ?array
+    private function getFilterData(ClassEntityCollection $classEntityCollection, string $className): ?array
     {
-        static $functionsData = [];
-        if (!array_key_exists($className, $functionsData)) {
-            $functions = $this->getAllUsedFunctions();
-            if (!isset($functions[$className])) {
-                return null;
-            }
-            $entity = $classEntityCollection->getEntityByClassName($className);
-            if (str_starts_with($entity->getFileName(), '/SelfDoc')) {
+        static $filtersData = [];
+        if (!array_key_exists($className, $filtersData)) {
+            $filters = $this->getAllUsedFilters();
+            if (!isset($filters[$className])) {
                 return null;
             }
 
-            $functionData['name'] = $functions[$className];
+            $functionData['name'] = $filters[$className];
+            $entity = $classEntityCollection->getEntityByClassName($className);
             $method = $entity->getMethodEntityCollection()->get('__invoke');
             $functionData['parameters'] = $method->getParameters();
-            $functionsData[$className] = $functionData;
+            $filtersData[$className] = $functionData;
         }
-        return $functionsData[$className];
+        return $filtersData[$className];
     }
 }
