+++ selfdoc/templates/_config.yml
