+++ selfdoc/templates/assets/doc_example.png
