+++ selfdoc/templates/assets/doc_structure.png
