+++ selfdoc/templates/assets/doc_structure2.png
