+++ selfdoc/templates/readme.md.twig
@@ -68,4 +68,4 @@ $docgen->generate();
 
 To update this documentation, run the following command:
 
-{{ 'php SelfDoc/docGen.php generate' | textToCodeBlock('console') }}
+{{ './bin/bumbleDocGen generate' | textToCodeBlock('console') }}
