+++ selfdoc/templates/tech/1.configuration/readme.md.twig
@@ -12,7 +12,7 @@ During the generation process, the data from these files is loaded into the <a>C
 
 Let's look at an example of a real configuration in more detail:
 
-{{ fileGetContents('%templates_dir%/../config.yaml') | textToCodeBlock('yaml') }}
+{{ fileGetContents('%WORKING_DIR%/bumble_doc_gen.yaml') | textToCodeBlock('yaml') }}
 
 In this example, we see the real configuration of the self-documentation of this project.
