+++ selfdoc/templates/tech/3.renderer/templatesDynamicBlocks.md.twig
