+++ selfdoc/templates/tech/4.pluginSystem/readme.md.twig
@@ -13,9 +13,9 @@ You can add your plugins to the configuration like this:
 
 ```yaml
 plugins:
-  - class: \SelfDoc\Configuration\Plugin\RoaveStubber\BetterReflectionStubberPlugin
-  - class: \SelfDoc\Configuration\Plugin\TwigFilterClassParser\TwigFilterClassParserPlugin
-  - class: \SelfDoc\Configuration\Plugin\TwigFunctionClassParser\TwigFunctionClassParserPlugin
+  - class: \SelfDocConfig\Plugin\RoaveStubber\BetterReflectionStubberPlugin
+  - class: \SelfDocConfig\Plugin\TwigFilterClassParser\TwigFilterClassParserPlugin
+  - class: \SelfDocConfig\Plugin\TwigFunctionClassParser\TwigFunctionClassParserPlugin
 ```
 
 {{ "Default plugins" | textToHeading('H2') }}
